+++ kedro_azureml/generator.py
@@ -154,9 +154,9 @@ class AzureMLPipelineGenerator:
             suffix = ":" + version
         return azureml_dataset_name + suffix
 
-    def _get_input_type(self, dataset_name: str, pipeline: Pipeline) -> Input:
+    def _get_input(self, dataset_name: str, pipeline: Pipeline) -> Input:
         if self._is_param_or_root_non_azureml_asset_dataset(dataset_name, pipeline):
-            return "string"
+            return Input(type="string")
         elif dataset_name in self.catalog.list() and isinstance(
             ds := self.catalog._get_dataset(dataset_name), AzureMLAssetDataSet
         ):
@@ -164,9 +164,22 @@ class AzureMLPipelineGenerator:
                 raise ValueError(
                     "AzureMLAssetDataSets with azureml_type 'uri_file' can only be used as pipeline inputs"
                 )
-            return ds._azureml_type
+            return Input(type=ds._azureml_type)
         else:
-            return "uri_folder"
+            return Input(type="uri_folder")
+
+    def _get_output(self, name):
+        if name in self.catalog.list() and isinstance(
+            ds := self.catalog._get_dataset(name), AzureMLAssetDataSet
+        ):
+            if ds._azureml_type == "uri_file":
+                raise ValueError(
+                    "AzureMLAssetDataSets with azureml_type 'uri_file' cannot be used as outputs"
+                )
+            # TODO: add versioning
+            return Output(type=ds._azureml_type, name=ds._azureml_dataset)
+        else:
+            return Output(type="uri_folder")
 
     def _from_params_or_value(
         self,
@@ -231,21 +244,11 @@ class AzureMLPipelineGenerator:
             },
             environment=self._resolve_azure_environment(),  # TODO: check whether Environment exists
             inputs={
-                self._sanitize_param_name(name): Input(
-                    type=self._get_input_type(name, pipeline)
-                )
+                self._sanitize_param_name(name): self._get_input(name, pipeline)
                 for name in node.inputs
             },
             outputs={
-                self._sanitize_param_name(name): (
-                    # TODO: add versioning
-                    Output(name=ds._azureml_dataset)
-                    if name in self.catalog.list()
-                    and isinstance(
-                        ds := self.catalog._get_dataset(name), AzureMLAssetDataSet
-                    )
-                    else Output()
-                )
+                self._sanitize_param_name(name): self._get_output(name)
                 for name in node.outputs
             },
             code=self.config.azure.code_directory,