+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 -   Added support for pydantic v2 and bumped minimal required pydantic version to `2.0.0` by [@froessler](https://github.com/fdroessler)
 -   Added ability to mark a node as deterministic (enables caching on Azure ML) by [@tomasvanpottelbergh](https://github.com/tomasvanpottelbergh)
 -   Explicitly disabled support for `AzureMLAssetDataSet` outputs of `uri_file` type by [@tomasvanpottelbergh](https://github.com/tomasvanpottelbergh)
+-   Made `AzureMLAssetDataSet` local and downloadable by default allowing their use in kedro sessions outside of pipeline runs e.g. `kedro ipython/jupyterlab` by [@froessler](https://github.com/fdroessler)
 -   Fixed FileNotFoundError for local runs (using `kedro run`) when using `AzureMLAssetDataSet` of type `uri_file` by [@Gabriel2409](https://github.com/Gabriel2409)
 
 ## [0.5.0] - 2023-08-11